The cncEnvIn function is no longer in the scope of a finish EDT.
The application can terminate as soon as all of the output data is
ready. The cncEnvOut function is still in a finish EDT, in case the user
wanted to create more EDTs to handle the graph output. However, it seems
that waiting until all output data is available is sufficent for
detecting the "end" of the graph computation.